% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{add_path_genes}
\alias{add_path_genes}
\title{Add Gene Names and logFC Values to goana Result}
\usage{
add_path_genes(go_res, top_table, gslist = NULL, species = "Hs", FDR = 0.05)
}
\arguments{
\item{go_res}{result of call to \code{\link{limma::goana}}}

\item{top_table}{result of limma::topTable with \code{n = Inf} with character column \code{ENTREZID} added. \code{topTable}
should be called with argument \code{n = Inf} to ensure that all genes with \code{adj.P.Val < FDR} are added}

\item{gslist}{result of \code{\link{get_gslist.go}}. If \code{NULL} (default) then function retrieved internally.}

\item{species}{Used by \code{get_gslist.go} if \code{gslist} is NULL.}

\item{FDR}{false discovery rate cutoff for differentially expressed genes. Numeric value between 0 and 1.}
}
\value{
\code{go_res} with columns \code{SYMBOL} and \code{logFC} containing lists of vectors with gene symbols and logFC
  values for genes with \code{adj.P.Val < FDR}.
}
\description{
Used to add gene name and logfc values to goana result prior to \code{\link{scrape_revigo}}. Allows \code{\link{r2d3_forcegraph}}
and other graphs to display gene names and logfc values for GO terms.
}
\examples{
## Not run:
## Example of setup assuming you have: y, design, and enids

library(limma)

fit <- lmFit(y, design)
fit <- eBayes(fit)

# add ENTREZID column to fit result (see geneid argument of ?limma::goana)
fit$genes <- data.frame(ENTREZID = enids)

# Standard GO analysis
go_res <- goana(fit, species="Hs")

# Differential expression analysis
top_table <- topTable(fit, n = Inf)

# example for upregulated ontologies
go_up <- go_res[go_res$P.Up < 10e-5 & go_res$P.Up < go_res$P.Down, ]
go_up <- add_path_genes(go_up, top_table)

# column ordering expected by scrape_revigo
go_up <- go_up[, c('P.Up', 'SYMBOL', 'logFC')]
data_dir <- tempdir()
scrape_revigo(data_dir, go_up)


}
