% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_forcegraph.R
\name{adjust_forcegraph_colors}
\alias{adjust_forcegraph_colors}
\title{Adjust forcegraph colors to compare multiple analyses}
\usage{
adjust_forcegraph_colors(data)
}
\arguments{
\item{data}{result of \code{append_forcegraph_annotations}}
}
\value{
\code{data} with colors adjusted.
}
\description{
Nodes for single analyses get hue 240(blue) and 0(red).
Nodes resulting from merged analyses get hue 275 (purple).
Saturation and lightness stay the same to indicate significance.
}
\keyword{internal}
