% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_forcegraph.R
\name{append_forcegraph_annotations}
\alias{append_forcegraph_annotations}
\title{Add gene names, logfc values, and analysis indicator to forcegraph data}
\usage{
append_forcegraph_annotations(data, data_dir)
}
\arguments{
\item{data}{result of \code{\link{convert_xgmml}}}

\item{data_dir}{directory with scraped revigo data}
}
\value{
\code{data} with columes merged_genes, logFC, and analysis, added to nodes data.frame
}
\description{
Used internally by \link{revigo_forcegraph}
}
\keyword{'cytoscape_map.xgmml')}
\keyword{<-}
\keyword{append_annotations(data,}
\keyword{convert_xgmml(xgmml_path)}
\keyword{data}
\keyword{data(go_up1)}
\keyword{data_dir}
\keyword{data_dir)}
\keyword{file.path(data_dir,}
\keyword{go_up1)}
\keyword{internal}
\keyword{scrape_revigo(data_dir,}
\keyword{tempdir()}
\keyword{xgmml_path}
