% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_revigo.R
\name{append_genes}
\alias{append_genes}
\title{Add gene names and logfc values to forcegraph data}
\usage{
append_genes(data, data_dir)
}
\arguments{
\item{data}{result of \code{\link{convert_xgmml}}}

\item{data_dir}{directory with scraped revigo data}
}
\value{
\code{data} with columes merged_genes and logfc added to nodes data.frame
}
\description{
Used internally by \link{r2d3_forcegraph}
}
\keyword{internal}
