% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{convert_xgmml}
\alias{convert_xgmml}
\title{Convert cytoscape xgmml to data.frames}
\usage{
convert_xgmml(xgmml_path)
}
\arguments{
\item{xgmml_path}{path to .xgmml file}
}
\value{
list of data.frames with nodes and links for forcegraph
}
\description{
Convert cytoscape xgmml to data.frames
}
