% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{data_to_json}
\alias{data_to_json}
\title{Format forcegraph data.frames to JSON}
\usage{
data_to_json(data)
}
\arguments{
\item{data}{result of \code{\link{convert_xgmml}}}
}
\value{
JSON objects
}
\description{
Format forcegraph data.frames to JSON
}
