% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_gs.names.go}
\alias{get_gs.names.go}
\title{Get names of gene set}
\usage{
get_gs.names.go(gslist.go, species = "Hs")
}
\arguments{
\item{species}{species identifier}

\item{gslist}{result of \code{\link{get_gslist}}}

\item{type}{either 'go' or 'kegg'}

\item{gs_dir}{Directory to save results to}
}
\value{
Description of \code{gslist} gene sets
}
\description{
Get names of gene set
}
\examples{
gslist.go <- get_gslist.go()
gs.names.go <- get_gs.names.go(gslist.go)

}
