% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{get_gslist.go}
\alias{get_gslist.go}
\title{Get GO geneset list}
\usage{
get_gslist.go(species = "Hs")
}
\arguments{
\item{species}{string to specify organism annotation package. Default \code{'Hs'} uses org.Hs.eg.db.}
}
\value{
Lists of named character vectors. Each list contains a character vector of ENTREZIDs with HGNC symbols as names
 for each GO id.
}
\description{
Used by \code{\link{add_path_genes}} so add significant genes to output of \code{limma::goana}.
Adapted from \code{limma::goana}
}
\examples{

gslist.go <- get_gslist.go('Hs')

}
\seealso{
\link[limma]{goana}, \link{add_path_genes}.
}
