% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_inconsistent_genes}
\alias{get_inconsistent_genes}
\title{Identify genes with inconsistent logFC values}
\usage{
get_inconsistent_genes(symbols, logfc)
}
\arguments{
\item{symbols}{character vector of gene names}

\item{logfc}{numeric vector of logFC values}
}
\value{
character vector of gene names where logFC values disagree
}
\description{
Used to resolve conflicts for two-analysis plots.
}
\keyword{internal}
