% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_forcegraph.R
\name{get_merged_annotations}
\alias{get_merged_annotations}
\title{Get merged gene names, logfc values, and analysis indicator}
\usage{
get_merged_annotations(data_dir)
}
\arguments{
\item{data_dir}{directory with scraped revigo data}
}
\value{
\code{tibble} with columes merged_genes, logFC, analysis, and id
}
\description{
Used internally by \link{revigo_forcegraph} and \code{revigo_scatterplot}
}
\keyword{'cytoscape_map.xgmml')}
\keyword{<-}
\keyword{convert_xgmml(xgmml_path)}
\keyword{data}
\keyword{data(go_up1)}
\keyword{data_dir}
\keyword{file.path(data_dir,}
\keyword{get_merged_annotations(data_dir)}
\keyword{go_up1)}
\keyword{internal}
\keyword{scrape_revigo(data_dir,}
\keyword{tempdir()}
\keyword{xgmml_path}
