% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplot.R
\name{heatplot}
\alias{heatplot}
\title{Plot heatmap for genes from goana/kegga analysis}
\usage{
heatplot(path_res)
}
\arguments{
\item{path_res}{data.frame, result of \code{limma::goana} or \code{limma::kegga} with
significant genes for each pathway added by \code{add_path_genes}.}
}
\value{
\code{ggplot} object.
}
\description{
modified from enrichplot
}
\examples{

# plot pathways merged by revigo under inflammatory response
data(go_up1)

data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
revigo_res <- read.csv(file.path(data_dir, 'rsc.csv'), row.names = 1, stringsAsFactors = FALSE)
go_infl <- row.names(revigo_res)[revigo_res$representative == 6954]
heatplot(go_up1[go_infl, ])

}
