% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_consistent}
\alias{is_consistent}
\title{Resolve conflicts of logfc for multi-sample analyses}
\usage{
is_consistent(symbols, exclude)
}
\arguments{
\item{symbols}{character vector of gene names}

\item{exclude}{gene names to exclude}
}
\value{
boolean of length \code{unlist(symbols)}
}
\description{
Resolve conflicts of logfc for multi-sample analyses
}
\keyword{internal}
