% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{revigo_scatterplot}
\alias{revigo_scatterplot}
\title{Revigo MDS plot}
\usage{
revigo_scatterplot(data_dir)
}
\arguments{
\item{data_dir}{directory with scraped revigo data}
}
\value{
ggplot object
}
\description{
Revigo MDS plot
}
\examples{

data(go_res)
data_dir <- tempdir()
scrape_revigo(data_dir, go_res)
revigo_scatterplot(data_dir)

}
