% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_treemap.R
\name{revigo_treemap}
\alias{revigo_treemap}
\title{Plot Revigo Treemap}
\usage{
revigo_treemap(data_dir, ...)
}
\arguments{
\item{data_dir}{Directory to store revigo scripts and output files}

\item{...}{additional arguments to \code{\link[treemap]{treemap}}}
}
\value{
treemap plot
}
\description{
Plot Revigo Treemap
}
\examples{

data(go_up1)
data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
revigo_treemap(data_dir, palette = "Accent", border.lwds = c(1,1), border.col = '#333333', fontcolor.labels = c('#333333', 'white'))

}
