% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevigoR.R
\docType{data}
\name{RevigoR-data}
\alias{RevigoR-data}
\alias{chart}
\alias{rd}
\alias{rdinput}
\title{RevigoR datasets}
\format{
An object of class \code{DAVIDFunctionalAnnotationChart} with 40 rows and 13 columns.

An object of class \code{RevigoRData} of length 2.

An object of class \code{data.frame} with 11 rows and 2 columns.
}
\usage{
data(chart)

data(rd)

data(rdinput)
}
\description{
Internally used data for RevigoR examples and vignettes.
}
\keyword{datasets}
