% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevigoR.R
\docType{package}
\name{RevigoR}
\alias{RevigoR}
\title{RevigoR}
\description{
RevigoR accesses the ReViGo service (http://revigo.irb.hr) for
given gene ontology terms and saves the data as a list like object of type
RevigoRData. Simple plotting functions are implemented, but can be extended 
upon if necessary.
}
\author{
Felix G M Ernst [aut]
}
