% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/RevigoRData-class.R
\name{getTableData}
\alias{getTableData}
\alias{getTreemapData}
\alias{RevigoRData-accessors}
\alias{getTableData,RevigoRData-method}
\alias{getTreemapData,RevigoRData-method}
\title{Accessing RevigoRData}
\usage{
getTableData(x)

getTreemapData(x)

\S4method{getTableData}{RevigoRData}(x)

\S4method{getTreemapData}{RevigoRData}(x)
}
\arguments{
\item{x}{a \code{\link[=RevigoRData-class]{RevigoRData}} object}
}
\value{
a DataFrame
}
\description{
\code{getTableData} and \code{getTreemapData} returns data from a
\code{RevigoRData} object.
}
\examples{
data(rd)
getTableData(rd)
getTreemapData(rd)
}
