% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/RevigoRData-class.R
\docType{class}
\name{RevigoRData-class}
\alias{RevigoRData-class}
\alias{RevigoRData}
\title{RevigoRData class}
\usage{
RevigoRData(table, treemap)
}
\arguments{
\item{table}{a DataFrame containing the table data of a RevigoR analysis}

\item{treemap}{a DataFrame containing the treemap data of a RevigoR analysis}
}
\description{
This class is a container for holding data returned from the ReViGo web 
service.
}
\section{Slots}{

\describe{
\item{\code{listData}}{a \code{\link[S4Vectors:SimpleList-class]{SimpleList}} with 
two entries, \code{table} and \code{treemap}}
}}

