% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevigoRData-class.R
\docType{methods}
\name{RevigoRData-functions}
\alias{RevigoRData-functions}
\alias{getTableData,RevigoRData-method}
\alias{getTreemapData}
\alias{getTreemapData,RevigoRData-method}
\title{Accessing RevigoRData}
\usage{
getTableData(x)

\S4method{getTableData}{RevigoRData}(x)

getTreemapData(x)

\S4method{getTreemapData}{RevigoRData}(x)
}
\arguments{
\item{x}{RevigoRData.}
}
\value{
DataFrame
}
\description{
\code{getTableData} and \code{getTreemapData} returns data froma
\code{RevigoRData} object.
}
\examples{
getTableData(rd)
getTreemapData(rd)
}
