% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/AllGenerics-export.R,
%   R/RevigoRData-class.R
\docType{class}
\name{RevigoRData}
\alias{RevigoRData}
\alias{getTableData}
\alias{getTreemapData}
\alias{getTableData,RevigoRData-method}
\alias{getTreemapData,RevigoRData-method}
\title{RevigoRData}
\usage{
getTableData(x)

getTreemapData(x)

RevigoRData(table, treemap)

\S4method{getTableData}{RevigoRData}(x)

\S4method{getTreemapData}{RevigoRData}(x)
}
\arguments{
\item{x}{RevigoRData.}

\item{table}{a DataFrame containting the table data of a RevigoR analysis}

\item{treemap}{a DataFrame containting the treemap data of a RevigoR analysis}
}
\value{
DataFrame
}
\description{
These classes are inheriting from List. It contains exactly two DataFrames, 
table and treemap

but require the presence of
the following columns to be valid.

\code{getTableData} and \code{getTreemapData} returns data froma
\code{RevigoRData} object.
}
\examples{
getTableData(rd)
getTreemapData(rd)
}
