% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_revigo.R
\name{scrape_revigo}
\alias{scrape_revigo}
\title{Submits GO result to revigo and downloads all scripts and outputs}
\usage{
scrape_revigo(data_dir, go_res)
}
\arguments{
\item{data_dir}{Directory to store revigo scripts and output files}

\item{go_res}{data.frame with row.names containing GO ids and first column p-values. An optional 3rd and 4th column
can contain lists of significant genes and their logFC values respectively (used for heatplots of GO terms).
An optional 5th column can contain a variable to indicate analysis identity (used to compare two seperate GO analyses).}
}
\description{
Note: requires installation of python packages through \code{\link{setup_env}}
}
\examples{

data(go_res)
data_dir <- tempdir()
scrape_revigo(data_dir, go_res)


}
