% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_env.R
\name{setup_env}
\alias{setup_env}
\title{Install python packages required by package}
\usage{
setup_env()
}
\value{

}
\description{
Creates virtualenv with name revigo and installs numpy, werkzeug, and robobrowser
}
