% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_forcegraph.R
\name{data_to_json}
\alias{data_to_json}
\title{Format forcegraph data.frames to JSON}
\usage{
data_to_json(data)
}
\arguments{
\item{data}{result of \code{\link{convert_xgmml}}}
}
\value{
JSON objects
}
\description{
Format forcegraph data.frames to JSON
}
\examples{
data(go_up1)
data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
xgmml_path <- file.path(data_dir, 'cytoscape_map.xgmml')
data <- convert_xgmml(xgmml_path)
data_to_json(data)

}
