% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_merged_annotations}
\alias{get_merged_annotations}
\title{Get merged gene names, logfc values, and analysis indicator}
\usage{
get_merged_annotations(data_dir)
}
\arguments{
\item{data_dir}{directory with scraped revigo data}
}
\value{
\code{tibble} with columes merged_genes, logFC, analysis, and id
}
\description{
Used internally by \link{revigo_forcegraph} and \code{revigo_scatterplot}
}
\examples{

data(go_up1)
data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
data <- get_merged_annotations(data_dir)

}
\keyword{internal}
