% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RevigoR-import.R
\name{getRevigo}
\alias{getRevigo}
\title{Get data from ReViGo}
\usage{
getRevigo(goTerms, revigoURL = "http://revigo.irb.hr", verbose = FALSE)
}
\arguments{
\item{goTerms}{a data.frame like object with two columns, "goTerms" and 
"p-value". If colnames are not given, the order is assumed.}

\item{revigoURL}{optional: the URL of the ReViGo service (default = 
\code{http://revigo.irb.hr})}

\item{verbose}{optional: whether to print more verbose information for the 
httr functions used to accessed the ReViGo service.}
}
\value{
RevigoRData
}
\description{
\code{getRevigo} access the ReViGo services with the given gene ontology 
terms.
}
\examples{
\donttest{
getRevigo()
}
}
