% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R,
%   R/RevigoR-RDAVIDWebService.R, R/RevigoR-import.R
\docType{data}
\name{getRevigo}
\alias{getRevigo}
\alias{getRevigo,DAVIDFunctionalAnnotationChart-method}
\alias{getRevigo}
\title{Get data from ReViGo}
\format{An object of class \code{NULL} of length 0.}
\usage{
\S4method{getRevigo}{DAVIDFunctionalAnnotationChart}(x, pValueCutoff = 0.05)

getRevigo(goTerms, cutoff = 0.7, isPValue = TRUE, whatIsBetter = "higher",
  orgID = 0, measure = "SIMREL", revigoURL = "http://revigo.irb.hr",
  verbose = FALSE)
}
\arguments{
\item{x}{a DAVIDFunctionalAnnotationChart object}

\item{goTerms}{a data.frame like object with two columns, "goTerms" and 
"p-value". If colnames are not given, the order is assumed.}

\item{revigoURL}{optional: the URL of the ReViGo service (default = 
\code{http://revigo.irb.hr})}

\item{verbose}{optional: whether to print more verbose information for the 
httr functions used to accessed the ReViGo service.}
}
\value{
a RevigoRData object

RevigoRData
}
\description{
Acc

\code{getRevigo} access the ReViGo services with the given gene ontology 
terms.
}
\examples{
\donttest{
go <- DAVIDFunctionalAnnotationChart()
getRevigo(go,
          pValueCutoff = 0.05)
}
\donttest{
getRevigo()
}
}
\keyword{datasets}
