% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics-export.R, R/RevigoR-plot.R
\docType{methods}
\name{plotRevigo}
\alias{plotRevigo}
\alias{plotRevigo}
\alias{plotRevigoScatter,RevigoRData-method}
\alias{plotRevigoTreemap,RevigoRData-method}
\title{Plotting ReViGo data}
\usage{
plotRevigoScatter(x)

plotRevigoTreemap(x)

\S4method{plotRevigoScatter}{RevigoRData}(x)

\S4method{plotRevigoTreemap}{RevigoRData}(x)
}
\arguments{
\item{x}{a RevigoRData object}
}
\value{
a ggplot2 plot
}
\description{
plotting stuff
}
\examples{
\donttest{
rd <- getRevigo()
plotRevigoScatter(rd)
plotRevigoTreemap(rd)
}
}
