% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_revigo.R
\name{r2d3_forcegraph}
\alias{r2d3_forcegraph}
\title{Generate forcegraph plot of revigo GO graph}
\usage{
r2d3_forcegraph(data_dir)
}
\arguments{
\item{data_dir}{Directory to store revigo scripts and output files}
}
\value{
r2d3 plot
}
\description{
If \link{add_path_genes} is called, hovering a node will show the logFC of significant genes in all GO terms
merged into the the representative GO term in question. At most 70 upregulated (red) and 70 downregulated (green)
genes with the largest absolute logFC are displayed.
}
\examples{

data(go_res)
data_dir <- tempdir()
scrape_revigo(data_dir, go_res)
r2d3_forcegraph(data_dir)


}
\seealso{
\link{add_path_genes} to enable heatplots on hover.
}
