% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_scatterplot.R
\name{revigo_scatterplot}
\alias{revigo_scatterplot}
\title{Revigo MDS plot}
\usage{
revigo_scatterplot(data_dir)
}
\arguments{
\item{data_dir}{directory with scraped revigo data}
}
\value{
ggplot object
}
\description{
Revigo MDS plot
}
\examples{


# single analysis
data(go_up1)
data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
revigo_scatterplot(data_dir)

# two analyses
data(go_up2)

go_up1$analysis <- 0
go_up2$analysis <- 1
go_up <- rbind(go_up1, go_up2)
scrape_revigo(data_dir, go_up)
revigo_scatterplot(data_dir)

}
