% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revigo_scatterplot.R
\name{revigo_scatterplot}
\alias{revigo_scatterplot}
\title{Revigo MDS plot}
\usage{
revigo_scatterplot(data_dir)
}
\arguments{
\item{data_dir}{directory with scraped revigo data}
}
\value{
r2d3 plot
}
\description{
If \link{add_path_genes} is called, hovering a node will show the logFC of significant genes in all GO terms
merged into the the representative GO term in question. At most 70 upregulated (red) and 70 downregulated (blue)
genes with the largest absolute logFC are displayed.
}
\details{
If \code{scrape_revigo} is used with two analyses (see examples), revigo ontolgies where no merge occured
across analyses will be shades of red and blue while ontologies where a merge occured across analyses
will be shades of pink For tooltip heatmaps with two analyses, any genes regulated in opposite directions
 are excluded.
}
\examples{


# single analysis
data(go_up1)
data_dir <- tempdir()
scrape_revigo(data_dir, go_up1)
revigo_scatterplot(data_dir)

# two analyses
data(go_up2)
data_dir <- tempdir()
go_up1$analysis <- 0
go_up2$analysis <- 1
go_up <- rbind(go_up1, go_up2)
scrape_revigo(data_dir, go_up)
revigo_scatterplot(data_dir)

}
