% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/RevigoRData-class.R
\docType{class}
\name{RevigoRData-class}
\alias{RevigoRData-class}
\alias{RevigoRData}
\title{RevigoRData}
\usage{
RevigoRData(table, treemap)
}
\description{
These classes are inheriting from List. It contains exactly two DataFrames, 
table and treemap

but require the presence of
the following columns to be valid.
}
\examples{
df <- DataFrame(a = 1, b = 2)
}
